Adding gtest to dev profile

+++ init/dev.sh
@@ -12,15 +12,21 @@ esac
 
 if [ ! -z "$go_suffix" ]
 then
-  go_dir=go${GO_VERSION}-linux-${go_suffix}
-  go_filename=go${GO_VERSION}.linux-${go_suffix}.tar.gz
-  go_url=https://dl.google.com/go/${go_filename}
-  tmp_dir=$(mktemp -d)
-  cd $tmp_dir && wget "$go_url" && tar xf $go_filename  && \
-    mv -vi go $SOFT_DIR/go-$GO_VERSION && \
-    cd $SOFT_DIR && ln -s go-$GO_VERSION go &&  \
-    rm -fr "${tmp_dir}"
-
+  if [ -d $SOFT_DIR/go-$GO_VERSION ]
+  e_header "Installing Go language go-$GO_VERSION"
+  then
+    e_arrow "An existing Go directory exists [$SOFT_DIR/go-$GO_VERSION]"
+  else
+    go_dir=go${GO_VERSION}-linux-${go_suffix}
+    go_filename=go${GO_VERSION}.linux-${go_suffix}.tar.gz
+    go_url=https://dl.google.com/go/${go_filename}
+    tmp_dir=$(mktemp -d)
+    cd $tmp_dir && wget -q "$go_url" && tar xf $go_filename  && \
+      mv -i go $SOFT_DIR/go-$GO_VERSION && \
+      cd $SOFT_DIR && ln -sf go-$GO_VERSION go &&  \
+      rm -fr "${tmp_dir}"
+    e_arrow "Go installaton done"
+  fi
 else
   e_warn "Unknwon architecture [$go_suffix]=> GO will not be installed"
 fi
